pyproject.toml
[project]
name = "ai-delivery-multiagent"
version = "0.1.0"
description = "Multi-agent software delivery generator for the IBM Gen AI 2026 hackathon."
readme = "README.md"
requires-python = ">=3.11,<3.13"
dependencies = [
  "fastapi>=0.115.0",
  "langchain>=0.3.0",
  "langgraph>=0.2.0",
  "pydantic>=2.8.0",
  "pyyaml>=6.0.1",
  "uvicorn>=0.30.0",
]

[project.optional-dependencies]
dev = [
  "pre-commit>=4.0.0",
  "pytest>=8.3.0",
  "ruff>=0.11.0",
]

[project.scripts]
ai-delivery = "ai_delivery.cli:main"

[build-system]
requires = ["hatchling>=1.25.0"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/ai_delivery"]

[tool.ruff]
line-length = 100
target-version = "py311"
src = ["src", "tests"]

[tool.ruff.lint]
select = [
  "E",
  "F",
  "I",
  "UP",
  "B",
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
addopts = "-ra"
